Crate bad_upwind

The bad_upwind crate provides a means to find out what results can occur when the upwind method is improperly used.

Section 1.3 of the book states that when solving the transport equation by the finite difference method, the upwind side of the advection must be used to calculate the value of the next time step in order to get a good answer.

We define the upwind method that uses information on the upwind side of the advection as a good upwind method and the one that uses information on the downwind side as a bad upwind method.

Both the good and bad upwind methods are implemented in this crate.

Using this crate, you can actually compute and see the difference between the good and bad upwind methods.

Modules

input

Module to read the input parameters.

output

Module to output the results.

upwind_solver

Solver for the transport equation using upwind method.

Functions

run

Run the solver and output the results.